Rename distributedWidth to responsiveWidth
